legacy-json-all, some cleanup
