+++ src/generators/legacy-json-all/index.mjs
@@ -1,7 +1,10 @@
 'use strict';
 
+import { writeFile } from 'fs/promises';
+import { join } from 'path';
+
 /**
- * @typedef {Array<TemplateValues>} Input
+ * @typedef {Array<import('../types.d.ts').Section[]>} Input
  *
  * @type {import('../types.d.ts').GeneratorMetadata<Input, import('./types.d.ts').Output>}
  */
@@ -15,5 +18,39 @@ export default {
 
   dependsOn: 'legacy-json',
 
-  async generate(input) {},
+  async generate(input, { output }) {
+    /**
+     * @type {import('./types.d.ts').Output}
+     */
+    const generatedValue = {
+      miscs: [],
+      modules: [],
+      classes: [],
+      globals: [],
+      methods: [],
+    };
+
+    for (const section of input) {
+      // Copy the relevant properties from each section into our output
+      for (const property of [
+        'miscs',
+        'modules',
+        'classes',
+        'globals',
+        'methods',
+      ]) {
+        if (section[property]) {
+          generatedValue[property].push(...section[property]);
+        }
+      }
+    }
+
+    await writeFile(
+      join(output, 'all.json'),
+      JSON.stringify(generatedValue),
+      'utf8'
+    );
+
+    return generatedValue;
+  },
 };
